+++ src/SubTask.java
@@ -1,13 +1,13 @@
 public class SubTask extends Task {
-    private final int epicID;
+    private final long epicID;
 
 
-    public SubTask(String title, String description, int taskId, Status status, int epicID) {
+    public SubTask(String title, String description, long taskId, Status status, long epicID) {
         super(title, description, taskId, status);
         this.epicID = epicID;
     }
 
-    public int getEpicID() {
+    public long getEpicID() {
         return epicID;
     }
 
@@ -21,7 +21,5 @@ public class SubTask extends Task {
                 ", epicID=" + epicID +
                 '}';
     }
-
-
 }
